# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: lacre
  version: "1"
  summary: Componente de facturación Veri*Factu para integrar en un ERP.
  description: |
    lacre genera los registros de facturación que exigen el Real Decreto 1007/2023 y la Orden
    HAC/1177/2024, los encadena con su huella y los remite a la AEAT. El ERP expide la factura;
    lacre solo registra.

    ## Tu primera factura en 15 minutos

    ### 1. Arranca lacre

    Necesita dos variables. Sin ninguna de las dos arranca:

    ```
    export LACRE_API_CLAVE=$(openssl rand -base64 32)
    export LACRE_QR_URL_BASE='https://prewww2.aeat.es/wlpl/TIKE-CONT/ValidarQR?'
    ```

    La segunda decide si los códigos QR que imprimas apuntan al entorno de pruebas o al de
    producción, que es `https://www2.agenciatributaria.gob.es/wlpl/TIKE-CONT/ValidarQR?`.

    ### 2. Da de alta al obligado tributario

    El obligado es quien expide la factura. Se da de alta por su NIF, y repetir la llamada no lo
    duplica:

    ```
    curl -X PUT http://localhost:8080/v1/obligados/89890001K \
      -H "Authorization: Bearer $LACRE_API_CLAVE" \
      -H 'Content-Type: application/json' \
      -d '{ "nombreRazon": "Obligado de prueba SL", "zonaHoraria": "Europe/Madrid" }'
    ```

    La zona horaria no es del despliegue sino de cada obligado: el huso de cada registro sale de
    ella, y Canarias (`Atlantic/Canary`) va una hora por detrás de la Península.

...
    remitirlas.

    ### 3. Registra la factura

    ```
...
      -d '{
...
        },
...

...

...

...
servers:
...
security:
  - claveDeApi: []
...
paths:
...
  /v1/obligados/{nif}:
    put:
      tags: [obligados]
      summary: Da de alta un obligado tributario o actualiza sus datos
      description: |
        Idempotente por NIF: crea el obligado si no existe (`201`) y, si ya existe, actualiza su
        nombre y su zona horaria (`200`). Por eso no pide `Idempotency-Key`.

        **Cambiar la zona horaria es seguro.** Cada registro guarda el huso con el que se
        calculó su huella, así que los anteriores no cambian: la zona nueva solo afecta a los
        que se generen después.

        No hay baja: un obligado con registros no se puede borrar, porque la norma obliga a
        conservarlos.
      operationId: altaDeObligado
      parameters:
        - name: nif
          in: path
          required: true
          description: NIF del obligado. Se valida su dígito de control.
          schema:
            type: string
            minLength: 9
            maxLength: 9
          example: "89890001K"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeticionObligado'
            examples:
              peninsula:
                summary: Obligado en la Península
                value:
                  nombreRazon: Obligado de prueba SL
                  zonaHoraria: Europe/Madrid
              canarias:
                summary: Obligado en Canarias
                value:
                  nombreRazon: Obligado canario SL
                  zonaHoraria: Atlantic/Canary
      responses:
        "200":
          description: El obligado ya existía; sus datos quedan actualizados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaObligado'
        "201":
          description: El obligado se ha dado de alta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespuestaObligado'
              example:
                nif: "89890001K"
                nombreRazon: Obligado de prueba SL
                zonaHoraria: Europe/Madrid
        "400":
          description: |
            El NIF no es válido, la zona horaria no existe o es un desfase fijo, o falta un campo.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problema'
              example:
                type: about:blank
                title: Datos de facturación no válidos
                status: 400
                detail: >-
                  La zona horaria +01:00 es un desfase fijo, sin horario de verano, y el huso de
                  los registros sería falso media parte del año. Use Europe/Madrid,
                  Atlantic/Canary o Africa/Ceuta
                codigo: validacion
        "401":
          $ref: '#/components/responses/SinCredencial'
...
components:
  securitySchemes:
    claveDeApi:
      type: http
      scheme: bearer
      description: |
        Una clave por despliegue, en `Authorization: Bearer <clave>`. lacre lo despliega el
        cliente en su propia infraestructura, así que un despliegue es un ERP y el ERP es la
        frontera de confianza. La clave se configura en `LACRE_API_CLAVE` y sin ella la
        aplicación no arranca.
...
  responses:
...
    SinCredencial:
      description: |
        Falta la cabecera `Authorization` o la clave no es la del despliegue. Las dos situaciones
        responden igual a propósito: distinguirlas convertiría el endpoint en un oráculo para
        quien esté probando claves.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problema'
          example:
            type: about:blank
            title: No autorizado
            status: 401
            detail: Falta o no es válida la clave de la API.
            codigo: no-autorizado
...
  schemas:
...
    PeticionObligado:
      type: object
      description: Datos de un obligado. El NIF identifica el recurso y va en la ruta.
      required: [nombreRazon, zonaHoraria]
      properties:
        nombreRazon:
          type: string
          maxLength: 120
          example: Obligado de prueba SL
        zonaHoraria:
          type: string
          description: |
            Identificador IANA de la zona desde la que se factura: `Europe/Madrid`,
            `Atlantic/Canary` o `Africa/Ceuta`. No se admiten desfases fijos como `+01:00` o
            `UTC`, que no tienen horario de verano y declararían un huso falso medio año.
          example: Europe/Madrid
    RespuestaObligado:
      type: object
      required: [nif, nombreRazon, zonaHoraria]
      properties:
        nif:
          type: string
          example: "89890001K"
        nombreRazon:
          type: string
        zonaHoraria:
          type: string
          example: Europe/Madrid
...
    Problema:
      type: object
      description: |
        Error en formato `ProblemDetail` (RFC 9457). El campo que se programa es `codigo`, que
        es estable; `title` y `detail` son para leer.
      required: [status, codigo]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
        codigo:
          type: string
          enum:
            - validacion
            - obligado-desconocido
            - registro-desconocido
            - clave-idempotencia-reutilizada
            - no-autorizado
        errores:
          type: array
          description: Solo en las violaciones de contrato; un elemento por campo.
          items:
            type: object
            required: [campo, mensaje]
            properties:
              campo:
                type: string
              mensaje:
                type: string